calculo do IMC
Calculei o IMC

+++ aula 3/exercicio2.py
@@ -1,4 +1,4 @@
-altura = int(input("Qual a sua altura em metros?"))
-peso = int(input("Qual o seu peso em quilos?"))
-IMC = peso/altura*altura
+altura = float(input("Qual a sua altura em metros?"))
+peso = float(input("Qual o seu peso em quilos?"))
+IMC = (peso/(altura*altura))
 print(f"A sua índice de massa corporal é:{IMC}")
